requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
openai
tavily-python
torch
python-dotenv
pydantic